app: error handling more updates to how errors are handled and returned also move validation to its own package and add a generic storage package to hold some errors

+++ pkg/storage/sqlite/accounts_delete.go
@@ -2,37 +2,66 @@ package sqlite
 
 import (
 	"context"
-	"database/sql"
+	"legocerthub-backend/pkg/storage"
 )
 
 // DeleteKey deletes a private key from the database
-func (storage *Storage) DeleteAccount(id int) error {
-	ctx, cancel := context.WithTimeout(context.Background(), storage.Timeout)
+func (store *Storage) DeleteAccount(id int) error {
+	ctx, cancel := context.WithTimeout(context.Background(), store.Timeout)
 	defer cancel()
 
+	tx, err := store.Db.BeginTx(ctx, nil)
+	if err != nil {
+		return err
+	}
+	defer tx.Rollback()
+
+	// check acct exists
+	// if scan in succeeds, key exists
 	query := `
+	SELECT id
+	FROM acme_accounts
+	WHERE id = $1
+	`
+
+	row := tx.QueryRowContext(ctx, query, id)
+	temp := -2
+	row.Scan(&temp)
+	if temp == -2 {
+		return storage.ErrNoRecord
+	}
+
+	// check not in use in certs
+	// if scan in succeeds, record exists in certificates
+	query = `
+	SELECT id
+	FROM certificates
+	WHERE acme_account_id = $1
+	`
+
+	row = tx.QueryRowContext(ctx, query, id)
+	temp = -2
+	row.Scan(&temp)
+	if temp != -2 {
+		return storage.ErrInUse
+	}
+
+	// delete
+	query = `
 	DELETE FROM
 		acme_accounts
 	WHERE
 		id = $1
 	`
 
-	// TODO: Ensure can't delete a key that is in use on an account or certificate
-	// this is already checked by FK constraint, but perhaps update this later.
-
-	result, err := storage.Db.ExecContext(ctx, query, id)
-	if err != nil {
-		return err
-	}
-	resultRows, err := result.RowsAffected()
+	_, err = store.Db.ExecContext(ctx, query, id)
 	if err != nil {
 		return err
 	}
 
-	// if the record wasn't found, return no rows error
-	// not exactly the purpose of this error, but it'll do
-	if resultRows == 0 {
-		return sql.ErrNoRows
+	err = tx.Commit()
+	if err != nil {
+		return err
 	}
 
 	return nil
